EcomDataVault (Data Science)/ML pipeline/requirements.txt
a2wsgi==1.10.7
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-fab==1.5.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.8.2
apispec==6.7.1
appdirs==1.4.4
appnope==0.1.4
argcomplete==3.5.1
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.35.63
botocore==1.35.63
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloud-sql-python-connector==1.13.0
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
configobj==5.0.9
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==43.0.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.44
dagshub-annotation-converter==0.1.2
databricks-sdk==0.37.0
dataclasses-json==0.6.7
datasets==3.1.0
db-dtypes==1.3.1
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.22.5
dvc==3.56.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
email_validator==2.2.0
ensure==1.0.4
entrypoints==0.4
executing==2.1.0
faiss-cpu==1.9.0
fastapi==0.115.5
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gql==3.5.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
groq==0.12.0
grpcio==1.67.1
grpcio-status==1.67.1
gto==1.7.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
iterative-telemetry==0.0.9
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kombu==5.4.2
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-huggingface==0.0.3
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langfuse==2.53.9
langgraph==0.2.48
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.143
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.17.2
mlflow-skinny==2.17.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.13.5
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.54.4
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
ordered-set==4.1.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
pendulum==3.0.0
pexpect==4.9.0
pg8000==1.31.2
pillow==11.0.0
platformdirs==3.11.0
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==5.1.0
pysbd==0.3.4
python-box==7.2.0
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.1.21
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.8
scramp==1.4.5
semver==3.0.2
sentence-transformers==3.3.0
sentry-sdk==2.18.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
time-machine==2.16.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
treelib==1.7.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
voluptuous==0.15.2
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==2.2.3
wheel==0.44.0
wirerope==0.4.8
wrapt==1.16.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.17.1
zc.lockfile==3.0.post1
zipp==3.21.0